fix: 修复扫码登录失败问题

+++ src/main/kotlin/online/bingiz/bilibili/video/internal/entity/QRCodeScanningData.kt
@@ -18,7 +18,7 @@ data class QRCodeScanningData(
     @SerializedName("refresh_token")
     val refreshToken: String,
     @SerializedName("timestamp")
-    val timestamp: Int,
+    val timestamp: Long,
     @SerializedName("code")
     val code: Int,
     @SerializedName("message")
